+++ test/helpers/mock_classes.dart
@@ -1,0 +1,48 @@
+import 'package:mocktail/mocktail.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:cloud_functions/cloud_functions.dart';
+import 'package:seasonbox/features/auth/data/auth_service.dart';
+import 'package:seasonbox/data/services/user_service.dart';
+import 'package:seasonbox/data/services/firestore_service.dart';
+import 'package:seasonbox/data/repositories/family_repository.dart';
+import 'package:seasonbox/data/services/biometric_service.dart';
+
+// ignore: subtype_of_sealed_class
+class MockFirebaseAuth extends Mock implements FirebaseAuth {}
+
+class MockUser extends Mock implements User {}
+
+class MockUserCredential extends Mock implements UserCredential {}
+
+// ignore: subtype_of_sealed_class
+class MockFirebaseFirestore extends Mock implements FirebaseFirestore {}
+
+// ignore: subtype_of_sealed_class
+class MockCollectionReference extends Mock
+    implements CollectionReference<Map<String, dynamic>> {}
+
+// ignore: subtype_of_sealed_class
+class MockDocumentReference extends Mock
+    implements DocumentReference<Map<String, dynamic>> {}
+
+// ignore: subtype_of_sealed_class
+class MockDocumentSnapshot extends Mock
+    implements DocumentSnapshot<Map<String, dynamic>> {}
+
+// ignore: subtype_of_sealed_class
+class MockFirebaseFunctions extends Mock implements FirebaseFunctions {}
+
+class MockHttpsCallable extends Mock implements HttpsCallable {}
+
+class MockHttpsCallableResult extends Mock implements HttpsCallableResult {}
+
+class MockAuthService extends Mock implements AuthService {}
+
+class MockUserService extends Mock implements UserService {}
+
+class MockFirestoreService extends Mock implements FirestoreService {}
+
+class MockFamilyRepository extends Mock implements FamilyRepository {}
+
+class MockBiometricService extends Mock implements BiometricService {}